added baby elf, boss, and enemy parent class
